improve output directory code logic
1. os.chdir() is a dangerous way to control file/directory access
in a script, as it is so easy to lose track (when writing) of
where the script should be. It is best avoided unless you really
need that behaviour. To refer to a file in a different directory,
it is safer to work with the relative file path.

2. os.mkdir() works, but is not the preferred option - os.makedirs()
is less prone to errors and race conditions.

3. it is better to define your output directory as a variable, and
operate on the contents of the variable, than write the directory
name several times in the script. This is defensive against typos,
and makes the code easier to read/understand.

See, e.g.

https://stackoverflow.com/questions/600268/mkdir-p-functionality-in-python