remove StringIO requirement for parsing UniProt output
The returned data from the UniProt search doesn't strictly require
StringIO. A useful principle to follow is not to introduce requirements
for extra modules/data format conversion if you don't have to, such as
when you can handle the data simply in the current script. Here, we
can do that.

The tasks that need to be done to process the UniProt output are:
1. split the returned data into lines, on the '\n' character
2. loop over the each of the (data) lines, and split them into tuples, by
   the whitespace that separates each accession from the CAZy families it
   is annotated with

That is now done with a single list comprehension, which removes the need
to import StringIO and convert a string into a StringHandle, just so it can
be looped over and processed in the same way.