modify UniProt FASTA query loop
The previous UniProt sequence query constructed an '.. OR .. OR ..'
query, which would fail silently if one or other accession was in error:
a faulty accession would not produce an error/warning if it was
not present. To avoid this, the loop now queries each accession singly,
and reports to the user if there is a problem with any sequence. In
general, it is good to avoid situations where queries/other activities
can fail silently without warning the user.

Also, the two-letter variable 'fa' and one-letter variable 'q' are
avoided. 'q' is avoided by not constructing the 'OR' query, and 'fa'
is avoided by having a new list named 'sequences', into which the
result of the UniProt query is written directly, without need for
an intermediate variable. The list of FASTA strings is then joined
in the `outfile.write()` call.

Finally, the path handling is modified so that the output filename is
defined for each family in a single line, with no need to redefine
`destination` every time the loop iterates. In general, anything that
is a static definition can sit outside the loop, otherwise it is an
inefficient use of processor time/cycles.